Switch from byte-oriented x25519 function to typed DH API
This introduces clear-on-drop semantics for X25519 secret keys. As a
side effect, it also causes these keys to be written in clamped form (as
x25519_dalek::StaticSecret stores the keys in clamped form internally).
Unclamped X25519 secret keys will still be read, but reading and then
writing the same key is no longer guaranteed to result in the same
encoding (and in any case, this is unnecessary for age use cases).