Avoid using String::split_off during armor reading
If a valid line is read, but the provided buffer can't hold it all, we
previously split the line at the buffer's length. This would panic if
the buffer's length did not match a UTF-8 boundary. Now we treat the
line itself as a byte buffer and track how much of it we have read.

Closes #60.