Give a more useful error when encryption fails with BrokenPipe
If a user tries to pipe to a program that is not reading from stdin (or
stops doing so early), the default Rust error is something like:

    Broken pipe (os error 32)

which is pretty opaque. While there might be other possible causes for
this error, we handle the most likely cause by wrapping this with a
custom error that suggests the user check whether the output is being
read. For example, this command would trigger the error:

    rage -p -a file.txt | cat foo

while this would not:

    rage -p -a file.txt | cat -

Closes str4d/rage#54.