Add support for OpenSSH keys encrypted with "aes256-cbc"
When OpenSSH introduced its current encrypted-key format, ssh-keygen
defaulted to "aes256-cbc" for the cipher. This was changed in OpenSSH 7.6
(released 2017-10-03) to "aes256-ctr", and is the same to this day.

Adding support for "aes256-cbc" ensures compatibility with all encrypted
keys generated by ssh-keygen in this format since it was introduced in
OpenSSH 6.5 (released 2014-01-30).

See also https://github.com/FiloSottile/age/issues/100