Update digest-related dependencies
We upgrade the hkdf, hmac, and rsa crates in lock-step because their
usages all depend on the sha2 crate, and upgrading from digest 0.8 to
0.9 is a breaking change that would otherwise require having two
versions of the sha2 crate.

We upgrade various other dependencies to ensure that we drop older
digest-related crate versions within our dependency tree where possible.

As a nice bonus, we can now depend on a published version of the rsa
crate that includes RSAES-OAEP.