Serialize stanzas with an explicit short last body line.
This enables a stream of stanzas to be parsed, such as in the upcoming
plugin IPC. Previously the parser could not distinguish an
stanza with a body of length 0 mod 64 from an incomplete stanza. This did
not cause issues for V1 header parsing because the recipients list is
always terminated with the MAC line, which has a distinct prefix.

V1 header parsing now accepts either kind of stanza body encoding.