age: Add age test vectors for multi-key attack mitigation
The test vectors pass. This was expected, as we already check the lengths
of encrypted file keys in all recipient types for which the attack was
mitigated.

See https://github.com/FiloSottile/age/commit/2194f6962c8bb3bca8a55f313d5b9302596b593b
for more details on the attack.