age: Cache V1 header bytes when reading to ensure MAC validation
Since str4d/rage#162 was merged, age files with both legacy and new stanza
bodies can now be parsed. However, age only serializes the new canonical
format. This caused headers with problematic legacy stanzas (with bodies
of length 0 mod 64) to be parsed correctly, but then incorrectly
serialized for MAC verification.

This commit caches the as-read encoding, to ensure the MAC is correctly
validated. We do not write out this cached header encoding, because we
never write headers that we've read, and we only generate headers for
writing without a cached encoding.