rage: Remove support for encrypting to HTTPS URLs
This added otherwise-unnecessary networking dependencies to `rage`, and
there are many decisions that need to be made when downloading a file
(e.g. what roots to trust?) that go beyond the APIs we want to focus on
here. Users should use a tool like `curl` or `wget` to download a
recipients file, and then pass it to `rage`.