Authenticate ciphertext length when seeking from end of StreamReader
`StreamReader::seek(SeekFrom::End(offset))` did not previously authenticate
the ciphertext length; if the ciphertext had been truncated or extended by
`adversary_offset`, it would instead seek to `offset + adversary_offset`.
This allowed an adversary with temporary control of an encrypted age file
to control the location of a plaintext read following a seek-from-end.

`age` now returns an error if the last chunk is invalid.

Fixes str4d/rage#195.