rage: Allow encrypting to identity files
This is specifically intended to support plugins that use symmetric
identities, where both encryption and decryption require a secret. We
don't want recipient strings (age1name1...) to contain secret values,
as this is easily overlooked by users. Instead, symmetric plugins would
only provide an identity file, and users would provide it during both
encryption and decryption.

As a side-effect, we now also support using identity files for native
age recipients and asymmetric plugins, as well as SSH keys; this keeps
the UX consistent.

To prevent ambiguity (e.g. if the user wants to decrypt but forgets the
-d/--decrypt flag), we add an -e/--encrypt flag and require it when
encrypting to identity files.