age: Defer owning Stanzas during header parsing until end
Headers are incrementally read by-byte, as we need to only read exactly
the header bytes. For headers with multiple stanzas, this previously
meant we were performing O(n!) heap allocations and drops during header
parsing (as each stanza in the header was incrementally read). Now, we
leave stanzas in borrowed form until we've finished reading the entire
header.