CI: Fix "Build age" interop job
The `master` branch of FiloSottile/age was renamed to `main`, which
didn't cause interop builds to break because (I think) GitHub was
redirecting branch requests. However, subsequently a placeholder
`master` branch was added for redirecting users of the old Homebrew
Tap, which stopped the redirect and caused the interop workflow to
attempt to build a branch that contains no code.