age-core: Reject invalid last lines in `AgeStanza` bodies
The length of a Base64 encoding can never be 1 mod 4, because that
only provides six of the eight bits necessary for encoding a byte.
Previously we checked that every line was valid Base64 (which works
because all lines except the last are 64 characters, which exactly
encodes 48 bytes).

In 0.7.0 we improved the parser efficiency by only running the Base64
decoder lazily. We replaced the per-line check with an `is_base64_char`
check, but forgot to reject the invalid subset of last-line lengths.