age: Allow ciphertexts that encrypt the empty plaintext
In str4d/rage#319 we required the last STREAM chunk to be non-empty, as
this is generally non-canonical (and can be represented instead by the
chunk marked as last being full). The sole exception to this is if the
plaintext has length zero, in which case there is no "previous" chunk,
and the single empty chunk is the canonical representation.

There generally isn't a good reason to create empty age ciphertexts, as
most identities don't provide authentication guarantees, but we've
allowed them in the past.