Read each identity file in a single pass
read_identities now performs a single pass over each identity file in
order to read it. Besides being perhaps a bit more efficient by reducing
the amount of I/O performed by rage, the main improvement that this
allows is that it's now possible to pass in the key from another program
or through some sort of keychain. For instance, it's now possible to
pass in a key using process substitution, i.e.

        rage -d -i <(...) encrypted_file.age

This also makes the behavior of rage a little more consistent with the
reference implementation (which allows this behavior).