ssh: Fix parsing of OpenSSH private key format
Support for AEAD ciphers was done in a way that means the `string` that
previously contained the entire ciphertext now only contains the
unauthenticated part of the ciphertext, and the tag is just appended
afterwards. This means the encryption header needs to be interpreted in
order to parse the full private key format.