age-plugin: Replace callback input `read_line_initial_text` with `write_str` and `read_line`
`read_line_initial_text` fills the tty input with the initial text. This
implies that the returned input is usually prefixed with the initial
text, or that the initial text can be removed.
This differs from `age` behaviour, which is to not allow the initial
text to be removed.

As a consequence, plugin leveraging callbacks for reading public input
(confirm, request_public_string) exhibit a different behaviour depending
on if they are called with `age` or `rage`.

This commit replaces `read_line_initial_text` with `write_str` and
`read_line`. The initial text is still displayed to the user, but it
does not affect the returned output.