rage: Parse identities before input file during decryption
Previously we avoided parsing `-i/--identity` flags before determining
whether the input file was passphrase-encrypted or not, as it is an
error to provide `-i/--identity` flags when decrypting a file encrypted
with a passphrase. However, once we support identities via standard
input (`-i -`), we won't be able to distinguish between standard input
containing an encrypted file and an encrypted identity, so instead of a
nice "stdin used multiple times" error we will get a misleading "invalid
header" error.

By parsing identities first, we ensure that the `-i/--identity` flags
are themselves used correctly, and then in the case that we wouldn't use
them due to passphrase encryption, we drop the parsed identities. It's a
bit more work on the unhappy path, but no more work on the happy path.